pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "assistant"
version = "0.1.0"
description = "a collection for my tools"
readme = "README.md"
authors = [{name = "PandaRyshan", email = "ryshan@pandas.run"}]
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
dependencies = [
    "Jinja2==3.1.2",
    "Werkzeug==2.2.3",
    "Flask==2.2.3",
    "Flask-SQLAlchemy==3.0.3",
    "Flask-Login==0.6.2",
    "Flask-WTF==1.1.1",
    "Flask-Admin==1.6.1",
    "Flask-Security==3.0.0",
    "Flask-Caching==2.0.2",
    "Flask-Mail==0.9.1",
    "email-validator==1.3.1",
    "azure-cognitiveservices-speech==1.26.0",
    "WTForms==3.0.1",
    "setuptools==65.5.1",
    "wheel==0.38.4",
    "itsdangerous==2.1.2",
    "python-dotenv==1.0.0"
]

[project.optional-dependencies]
test = ["pytest==7.2.2"]
prod = ["gunicorn==20.1.0"]

[tool.pytest]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
